V3.72 (#358)
* updating SEMVER to v3.72.1

* adding new branch to CI

* adding a tensor serialization format skeleton

* standardize reporting formats among cfloat and lns

* standardized reporting format for integer<>

* standardize reporting format for fixpnt

* standardize reporting format for posit

* standardize reporting format for dbns

* WIP: vector serialization with Scalar type specialization skeleton

* WIP: a serialization approach of linear algebra collections

* WIP: data file structure and typeIds

* WIP: adding type identifiers and template parameter sets

* bug fix: need to declar to_hex inline to avoid multiple object definitions during linking

* enabling long double to_hex in gcc

* WIP: streamlining the manipulators in native floating-point

* rearchitecting the native IEEE-754 constexpr/nonconstexpr setup

* adding nonconstexpr BitCast to native manipulators.hpp

* renaming adaptive precision decimal to edecimal to make room for fixed-sized decimal integer class

* renaming adaptive precision decimal integer to elastic decimal

* WIP: storing renaming work

* renaming rational to elastic rational

* WIP: checkpointing edits

* improving code quality by moving 32bit into 64bit math

* fixing include hierarchy for gcc long double

* include hierarchy for native types fix for clang

* quick fix for missing type_tag in gcc build

* fix for clang on x86: didn't have a long double to_hex

* new clang15 and clang16 build containers

* bug fix of is_integer<> traits

* quantization benchmark for fp8/fp12

* refining the mixed-precision FMA experiment to demonstrate equivalence to FP64

* bug fix of stddev

* checkpointing the FMA quantization experiment

* save/restore for an ieee-754 vector/matrix

* save/restore for a cfloat, but a fixed one

* rearchitecting the FDP overloads in BLAS

* enabling lns conversion regression testing

* adding assignment and conversion skeletons to dbns

* adding assignment to native ieee-754 to dbns

* adding arithmetic regression tests for dbns and changing to explicit multi-limb data structure

* code hygiene for MSVC

* perminantly disabling dbns configurations that do not have two exponent fields

* WIP: debugging the approximation of unity rounding approach

* adding dbns rounding regression test

* compilation fix for gcc and clang

* experiments with dbns rounding

* adding arithmetic event statistics to LNS and DBNS types to trace rounding

* WIP: custom dot product quantization experiments

* adding summary statistics to double check random distributions

* streamlining the quantization benchmarking code

* disabling native DBNS arithmetic on multiply and divide: it is now fully marshalling through double arithmetic

* adding different Pareto generating dot product experiments

* adding regression skeletons for saturating cfloats

* passing regressions of all to_blocktriple suites for saturated cfloats

* passing regressions of from_blocktriple

* adding a circuit complexity measure to quantization experiment generator

* circuit complexity generators per arithmetic type

* enumerating ALUs

* FMA ALU complexity experiment

+++ include/universal/number/posit/exponent.hpp
@@ -64,7 +64,7 @@ public:
 			if (msb >= 0 && es > 0) {
 				nrExponentBits = (static_cast<unsigned>(msb) >= es - 1ull ? es : static_cast<unsigned>(msb) + 1ull);
 				for (unsigned i = 0; i < nrExponentBits; i++) {
-					_exp[es - 1 - i] = _raw_bits[msb - i];
+					_exp[static_cast<uint64_t>(es) - 1ull - i] = _raw_bits[static_cast<uint64_t>(msb) - i];
 				}
 			}
 			set(_exp, nrExponentBits);
@@ -103,7 +103,7 @@ public:
 		_NrOfBits = (nbits - 1 - nrRegimeBits > es ? es : nbits - 1 - nrRegimeBits);
 		if (_NrOfBits > 0) {
 			if (_NrOfBits < es) {
-				rounding_mode = _Bits[es - 1 - _NrOfBits] ? GEOMETRIC_ROUND_UP : GEOMETRIC_ROUND_DOWN; // check the next bit to see if we need to geometric round
+				rounding_mode = _Bits[static_cast<uint64_t>(es) - 1ull - _NrOfBits] ? GEOMETRIC_ROUND_UP : GEOMETRIC_ROUND_DOWN; // check the next bit to see if we need to geometric round
 				if (_trace_rounding) std::cout << "truncated exp" << (rounding_mode == GEOMETRIC_ROUND_UP ? " geo-up " : " geo-dw ");
 			}
 			else {
@@ -111,7 +111,6 @@ public:
 					// use the fraction to determine rounding as this posit has fraction bits
 					rounding_mode = ARITHMETIC_ROUNDING;
 					if (_trace_rounding) std::cout << "arithmetic  rounding ";
-
 				}
 				else {
 					// this posit is in the geometric regime and has consumed all the bits
